requirements.txt
appdirs==1.4.3
astroid==2.3.3
autopep8==1.5
Babel==2.8.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.5.0
importlib-resources==1.3.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.2
oauthlib==3.1.0
pathspec==0.8.1
phonenumbers==8.11.5
psycopg2==2.8.4
psycopg2-binary==2.8.6
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.4.3
pylint-plugin-utils==0.6
python3-openid==3.2.0
pytz==2019.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.14.0
social-auth-core==4.1.0
sqlparse==0.3.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.0.10
whitenoise==3.3.1
wrapt==1.11.2
zipp==3.1.0
